So the idea is this
I created a struct and this struct holds N and value. Ill have to change the N to RN but ill do it after my break.

So starting in main, I first call collatz_wrapper. What the wrapper does is it create a cache to store the values that I get from collatz and then generates N random numbers and and calculates the amount of steps it takes to do collatz. Then it takes those steps, it takes the current RN, and the cache, and its size and it sends it to input. Wait i just realized something. The reason why my program is returning nothing is because I am saying cacheSize  < 20 but cacheSize is 20. So I will say cacheIndex < cacheSize instead and i will inc cache Index. But anyways, I convert RN to char and I set the the char N to to key which is RN. Then I do the same for value but this time i set steps to it. Then I set its key and steps to their corresponding values then I return cache. Ill explain it after my break. Pce