debugged my code in online gdb and LRU works now
So here is what I did
Well first of all, I changed the value in the struct to int because %c in the csv was printing nothing.

So what I saw after running a lot of tests was that if N < CacheSize, caching is not needed. However if N > CacheSize then Caching is needed

So I have a counter in input called cacheIndex and once this number >= the cacheSize, I call LRU and replace the least recently used item.  Now I do >= because the cache is full when it equals cacheSize so we got to start doing LRU.