did a lot of things
Added a header file to organize my code

Upped up my registers to 64 bit.

Added a csv but it is not an actual file, its just a printf

Added a method for calculating mean time but it dosent work as intended and i dont know why

Going to add cache hit but i need to really understand what it means.